+++ govtool/backend/sql/list-dreps.sql
@@ -1,8 +1,18 @@
+WITH DRepDistr AS (
+  SELECT
+    *,
+    ROW_NUMBER() OVER(PARTITION BY drep_hash.id ORDER BY drep_distr.epoch_no DESC) AS rn
+  FROM drep_distr
+  JOIN drep_hash
+  on drep_hash.id = drep_distr.hash_id
+)
+
 SELECT
    encode(dh.raw, 'hex'),
    va.url,
    encode(va.data_hash, 'hex'),
-   dr_deposit.deposit
+   dr_deposit.deposit,
+   DRepDistr.amount
 FROM drep_hash dh
 JOIN (
   SELECT dr.id, dr.drep_hash_id, dr.deposit,
@@ -18,4 +28,5 @@ LEFT JOIN (
 ) as dr_voting_anchor
 on dr_voting_anchor.drep_hash_id = dh.id and dr_voting_anchor.rn = 1
 left JOIN voting_anchor va ON va.id = dr_voting_anchor.voting_anchor_id
-
+LEFT JOIN DRepDistr
+on DRepDistr.hash_id = dh.id and DRepDistr.rn = 1